Create Shadow_Blade.js
This macro creates an item called Shadow Blade and a spell called Shadow Booming Blade, they leave a dae effect on a target when it is hit. It disappears when the target moves or when its the source actors turn. I couldn't automate the damage dealing part when they move so just use other damage roll on the target that moves.